requirements.txt
streamlit
transformers
torch
pillow
einops
